move usePaymentMethodsQuery in a sharable hook
