+++ src/components/customers/paymentMethodsList/DeletePaymentMethodDialog.tsx
@@ -1,9 +1,9 @@
 import { forwardRef, useImperativeHandle, useRef, useState } from 'react'
 
-import { PaymentMethodItem } from '~/components/customers/paymentMethodsList/types'
 import { WarningDialog, WarningDialogRef } from '~/components/WarningDialog'
 import { DestroyPaymentMethodInput } from '~/generated/graphql'
 import { useInternationalization } from '~/hooks/core/useInternationalization'
+import { PaymentMethodItem } from '~/hooks/customer/usePaymentMethodsList'
 
 interface DeletePaymentMethodDialogProps {
   paymentMethod?: PaymentMethodItem
