--- src/components/customers/paymentMethodsList/types.ts
@@ -1,3 +1,0 @@
-import { PaymentMethodsQuery } from '~/generated/graphql'
-
-export type PaymentMethodItem = PaymentMethodsQuery['paymentMethods']['collection'][number]
